change update and trye pagination.

+++ index.php
@@ -1,12 +1,15 @@
 <!DOCTYPE html>
 <html lang="en">
+
 <head>
     <link rel="stylesheet" href="style.css">
     <meta charset="UTF-8">
     <meta name="viewport" content="width=device-width, initial-scale=1.0">
-    <title>Document</title>
+    <title>Historique</title>
 </head>
+
 <body>
+
 
     <!-------------------- HEADER -------------------->
     <header>
@@ -17,7 +20,7 @@
             </ul>
         </div>
     </header>
-    
+
     <!-------------------- MAIN -------------------->
     <main>
         <section>
@@ -28,32 +31,41 @@
                 <article>Prix</article>
             </div>
 
-    <!-------------------- PHP -------------------->
+            <!-------------------- PHP SELECT QUERY -------------------->
             <?php
-                try{$bdd = new PDO('mysql:host=localhost;dbname=ampoule','root');
-                }
-                catch(PDOException $e){
-                    echo 'Echec de la connexion : ' .$e->getMessage();
-                }
+            try {
+                $bdd = new PDO('mysql:host=localhost;dbname=ampoule', 'root');
+            } catch (PDOException $e) {
+                echo 'Echec de la connexion : ' . $e->getMessage();
+            }
+            $table = $bdd->query('SELECT * FROM gestion ORDER BY date_change LIMIT 0,5');
+            foreach ($table as $donnee) {
+                $id_gestion = htmlspecialchars($donnee['id_gestion']);
+                $date_change = htmlspecialchars($donnee['date_change']);
+                $floor = htmlspecialchars($donnee['floor']);
+                $position = htmlspecialchars($donnee['position']);
+                $price = htmlspecialchars($donnee['price']);
 
-                $table = $bdd->query('SELECT id_gestion, date_change, floor, position, price FROM gestion');
-                foreach($table as $donnee){
-                    $id_gestion= $donnee['id_gestion'];
-                    $date_change= $donnee['date_change'];
-                    $floor= $donnee['floor'];
-                    $position= $donnee['position'];
-                    $price= $donnee['price'];
-
-                    echo '<div><article> '.$date_change.'</article>
-                               <article>Etage n°'.$floor.'</article>
-                               <article>Côté '.$position.'</article>
-                               <article> '.$price.'€</article>
-                                    <a href="http://localhost/Ampoule/edit.php?id_gestion='.$id_gestion.'&floor='.$floor.'"><img src="../Ampoule/edit.svg" alt="" class="image"></a>
-                                    <a href="http://localhost/Ampoule/delete.php?id_gestion='.$id_gestion.'"><img src="../Ampoule/poubelle.svg" alt="" class="image"></a>
+                echo '<div><article> ' . $date_change . '</article>
+                               <article>Etage n°' . $floor . '</article>
+                               <article>Côté ' . $position . '</article>
+                               <article> ' . $price . '€</article>
+                                    <a href="http://localhost/Ampoule/edit.php?id_gestion=' . $id_gestion . '&date_change=' . $date_change . '&floor=' . $floor . '&position=' . $position . '&price=' . $price . '"><img src="../Ampoule/edit.svg" alt="" class="image"></a>
+                                    <a href="http://localhost/Ampoule/delete.php?id_gestion=' . $id_gestion . '" onclick="return confirm(\'Etes-vous sûr de vouloir supprimer cette entrée ?\');"><img src="../Ampoule/poubelle.svg" alt="" class="image"></a>
                           </div>';
-                }
+            }
             ?>
+            <div class="pagination">
+                <a href="#"><< Previous</a>
+                <a href="#" title="Algorithm">1</a>
+                <a href="#" title="DataStructure">2</a>
+                <a href="#" title="Languages">3</a>
+                <a href="#" title="Interview" class="active">4</a>
+                <a href="#" title="practice">5</a>
+                <a href="#">Next >></a>
+            </div>
         </section>
     </main>
 </body>
+
 </html>
